Desafio2 terminado, excepto README

+++ Proyecto/Desafio2/src/main/java/sv/edu/udb/desafio1/service/BookServiceImpl.java
@@ -1,8 +1,12 @@
 package sv.edu.udb.desafio1.service;
 
+import sv.edu.udb.desafio1.dto.BookDTO;
+import sv.edu.udb.desafio1.exception.ResourceNotFoundException;
 import sv.edu.udb.desafio1.model.Book;
 import sv.edu.udb.desafio1.repository.BookRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,7 +23,26 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Book createBook(Book book) {
+    public Page<Book> getAllBooksPaged(Pageable pageable) {
+        return bookRepository.findAll(pageable);
+    }
+
+    @Override
+    public Book createBook(BookDTO bookDTO) {
+        Book book = new Book();
+        book.setTitle(bookDTO.getTitle());
+        book.setAuthor(bookDTO.getAuthor());
+        book.setPublicationYear(bookDTO.getPublicationYear());
+        return bookRepository.save(book);
+    }
+
+    @Override
+    public Book updateBook(Long id, BookDTO bookDTO) {
+        Book book = bookRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Book not found with id " + id));
+        book.setTitle(bookDTO.getTitle());
+        book.setAuthor(bookDTO.getAuthor());
+        book.setPublicationYear(bookDTO.getPublicationYear());
         return bookRepository.save(book);
     }
 
@@ -30,7 +53,15 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteBook(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Book not found with id " + id);
+        }
         bookRepository.deleteById(id);
     }
 
+    @Override
+    public Book getBookById(Long id) {
+        return bookRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Book not found with id " + id));
+    }
 }
